Adds comparison to CC manifold for fci wave function
Add the method to the fci wave function to check whether
the CC manifold curves towards the fci wave function.

More tests are still needed. Not properly working for the
CCSD manifold

+++ src/test/unit/test_wf_int_norm.py
@@ -31,16 +31,16 @@ class He2StringIndicesTestCase(unittest.TestCase):
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[0].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
                         self.assertEqual(Index[8].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                 elif i == 1:
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[0].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                         self.assertEqual(Index[8].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
             else:
                 self.fail(msg='Too many indices!')
         irrep = 4
@@ -54,16 +54,16 @@ class He2StringIndicesTestCase(unittest.TestCase):
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[4].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
                         self.assertEqual(Index[12].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                 elif i == 1:
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[4].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                         self.assertEqual(Index[12].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
             else:
                 self.fail(msg='Too many indices!')
 
@@ -79,16 +79,16 @@ class He2StringIndicesTestCase(unittest.TestCase):
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[0].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
                         self.assertEqual(Index[8].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                 elif i == 1:
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[0].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                         self.assertEqual(Index[8].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
             else:
                 self.fail(msg='Too many indices!')
         irrep = 4
@@ -102,16 +102,16 @@ class He2StringIndicesTestCase(unittest.TestCase):
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[4].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
                         self.assertEqual(Index[12].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                 elif i == 1:
                     with self.subTest('He2, occ of singles',
                                       i=i, irrep=irrep):
                         self.assertEqual(Index[4].occ_orb,
-                                         np.array([0], dtype=np.int8))
+                                         np.array([0], dtype=np.intc))
                         self.assertEqual(Index[12].occ_orb,
-                                         np.array([1], dtype=np.int8))
+                                         np.array([1], dtype=np.intc))
             else:
                 self.fail(msg='Too many indices!')
 
@@ -127,9 +127,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                 self.assertAlmostEqual(Index.C, -0.02677499 / self.He2_wf.norm)
                 if i == 0:
                     self.assertEqual(Index[0].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
                     self.assertEqual(Index[8].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -144,9 +144,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                 self.assertAlmostEqual(Index.C, -0.02720985 / self.He2_wf.norm)
                 if i == 0:
                     self.assertEqual(Index[4].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
                     self.assertEqual(Index[12].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -161,9 +161,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                 self.assertAlmostEqual(Index.C, -0.03191829 / self.He2_wf.norm)
                 if i_Ind == 0:
                     self.assertEqual(Index[0].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
                     self.assertEqual(Index[8].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -178,9 +178,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                 self.assertAlmostEqual(Index.C, -0.05013654 / self.He2_wf.norm)
                 if i_Ind == 0:
                     self.assertEqual(Index[4].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
                     self.assertEqual(Index[12].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
 
@@ -198,9 +198,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                        / self.He2_CCwf.norm)
                 if i == 0:
                     self.assertEqual(Index[0].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
                     self.assertEqual(Index[8].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -216,9 +216,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                        -0.02730136 / self.He2_CCwf.norm)
                 if i == 0:
                     self.assertEqual(Index[4].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
                     self.assertEqual(Index[12].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -234,9 +234,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                        -0.03210796 / self.He2_CCwf.norm)
                 if i_Ind == 0:
                     self.assertEqual(Index[0].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
                     self.assertEqual(Index[8].occ_orb,
-                                     np.array([0, 1], dtype=np.int8))
+                                     np.array([0, 1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         # ----------
@@ -253,9 +253,9 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                        / self.He2_CCwf.norm)
                 if i_Ind == 0:
                     self.assertEqual(Index[4].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
                     self.assertEqual(Index[12].occ_orb,
-                                     np.array([1], dtype=np.int8))
+                                     np.array([1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
 
@@ -350,8 +350,8 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                        / self.He2_CCwf.norm)
 
     def test_make_occ_ind_doub_He2(self):
-        zero_arr_1 = np.array([0], dtype=np.int8)
-        one_arr_1 = np.array([1], dtype=np.int8)
+        zero_arr_1 = np.array([0], dtype=np.intc)
+        one_arr_1 = np.array([1], dtype=np.intc)
         i = j = 0
         irrep_i = irrep_j = 0
         irrep_a = irrep_b = 0
@@ -495,10 +495,10 @@ class He2StringIndicesTestCase(unittest.TestCase):
         for i_Ind, Index in enumerate(self.He2_wf._string_indices_spirrep(0)):
             if i_Ind == 0:
                 self.assertEqual(Index.occ_orb,
-                                 np.array([0], dtype=np.int8))
+                                 np.array([0], dtype=np.intc))
             elif i_Ind == 1:
                 self.assertEqual(Index.occ_orb,
-                                 np.array([1], dtype=np.int8))
+                                 np.array([1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
         for i_Ind, Index in enumerate(self.He2_wf._string_indices_spirrep(
@@ -506,7 +506,7 @@ class He2StringIndicesTestCase(unittest.TestCase):
                                                occ_type='R'))):
             if i_Ind == 0:
                 self.assertEqual(Index.occ_orb,
-                                 np.array([0, 1], dtype=np.int8))
+                                 np.array([0, 1], dtype=np.intc))
             else:
                 self.assertTrue(False, msg='Too many indices!')
